Rename Custom User Service to indicate that its main role is parsing user data from security context
