+++ api-gateway/src/main/java/com/drbsimon/apigateway/controller/filter/PreFilter.java
@@ -1,10 +1,7 @@
 package com.drbsimon.apigateway.controller.filter;
 
 import com.drbsimon.apigateway.model.entity.Visitor;
-import com.drbsimon.apigateway.security.CustomUserDetailsService;
-import com.drbsimon.apigateway.security.JwtTokenFilter;
-import com.drbsimon.apigateway.security.service.JwtTokenServices;
-import com.drbsimon.apigateway.service.dao.VisitorDao;
+import com.drbsimon.apigateway.security.service.ParseVisitorSecurityService;
 import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.context.RequestContext;
 import com.netflix.zuul.exception.ZuulException;
@@ -14,7 +11,7 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class PreFilter extends ZuulFilter {
-    private final CustomUserDetailsService customUserDetailsService;
+    private final ParseVisitorSecurityService parseVisitorSecurityService;
 
     @Override
     public String filterType() {
@@ -34,7 +31,7 @@ public class PreFilter extends ZuulFilter {
     @Override
     public Object run() throws ZuulException {
         RequestContext ctx = RequestContext.getCurrentContext();
-        Visitor visitor = customUserDetailsService.parseVisitorFromToken();
+        Visitor visitor = parseVisitorSecurityService.parseVisitorFromToken();
         String foundUserId = (visitor != null) ?  visitor.getId().toString() : null;
         ctx.addZuulRequestHeader("userid", foundUserId);
         return null;
